+++ src/app/api/domains/[domain]/[cluster]/workflows/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from 'next/server';
 
-import type { RouteParams } from '@/handlers/list-workflows/list-workflows.types';
-import { listWorkflows } from '@/handlers/list-workflows/list-workflows';
+import type { RouteParams } from '@/server/route-handlers/list-workflows/list-workflows.types';
+import { listWorkflows } from '@/server/route-handlers/list-workflows/list-workflows';
 
 export async function GET(
   request: NextRequest,
